backend/requirements.txt
fastapi
uvicorn
torch
torchio
torchvision
pillow
numpy
matplotlib
torchvision
